Update Aug 28th

+++ resources/views/characters/index.blade.php
@@ -19,7 +19,7 @@
                     <th>Offspec</th>
                     <th>Item Level</th>
                     <th>Wowprogress</th>
-                    <th>is Main</th>
+                    @can('Character Main Status')<th>is Main</th>@endcan
                     <th>Last Update</th>
                     <th>Actions</th>
                 </tr>
@@ -34,35 +34,47 @@
                         <td>{{ $character->gear }}</td>
                         <td><a href="{{ $character->wowprogress }}" target="_blank">Open</a></td>
                         <td>
-                            {{ Form::open(['method' => 'POST', 'class' =>'form-inline', 'route' => ['characters.status', $character->id] ]) }}
-                            {{ Form::select('main', [0 => 'No', '1' => 'Yes'], $character->main , array('class' => 'form-control-sm')) }}
-                            <div class="input-bar-item">
-                                <button class="btn btn-light"><i class="fab fa-accusoft"></i></button>
-                            </div>
-                            {!! Form::close() !!}
+                            @can('Character Main Status')
+                                {{ Form::open(['method' => 'POST', 'class' =>'form-inline', 'route' => ['characters.status', $character->id] ]) }}
+                                {{ Form::select('main', [0 => 'No', '1' => 'Yes'], $character->main , array('class' => 'form-control-sm')) }}
+                                <div class="input-bar-item">
+                                    <button class="btn btn-light"><i class="fab fa-accusoft"></i></button>
+                                </div>
+                                {!! Form::close() !!}
+                            @endcan
                         </td>
                         <?php $dateTemp = DateTime::createFromFormat('Y-m-d H:i:s', $character->updated_at);
                         $dateTemp = $dateTemp->format('d M Y H:i');
                         ?>
                         <td>{{$dateTemp}}</td>
                         <td>
-                            {!! Form::open(['method' => 'DELETE', 'class' =>'form-inline', 'route' => ['characters.destroy', $character->id] ]) !!}
-                            <div class="form-group">
-                                <a href="{{ route('characters.edit', $character->id) }}" ><i class="fas fa-edit fa-lg"></i></a>
-                            </div>
-                            <div class="form-group">
-                                <div class="input-bar-item">
-                                    <button class="btn btn-light"><i class="fas fa-trash fa-lg"></i></button>
+                            @can('Character Delete')
+                                {!! Form::open(['method' => 'DELETE', 'class' =>'form-inline', 'route' => ['characters.destroy', $character->id] ]) !!}
+                            @endcan
+                            @can('Character Edit')
+                                <div class="form-group">
+                                    <a href="{{ route('characters.edit', $character->id) }}" ><i class="fas fa-edit fa-lg"></i></a>
                                 </div>
-                            </div>
+                            @endcan
+                            @can('Character Delete')
+                                <div class="form-group">
+                                    <div class="input-bar-item">
+                                        <button class="btn btn-light"><i class="fas fa-trash fa-lg"></i></button>
+                                    </div>
+                                </div>
                             {!! Form::close() !!}
+                            @endcan
                         </td>
                     </tr>
                 @endforeach
                 </tbody>
             </table>
         </div>
-        <br><a href="{{ route('characters.create') }}" class="btn btn-success">Add Character</a></br>
+        <br>
+        @can('Character Create')
+            <a href="{{ route('characters.create') }}" class="btn btn-success">Add Character</a></br>
+        @endcan
+
     </div>
 @endsection
 @section('javascript')
@@ -75,11 +87,5 @@
                 order: [ 1, 'asc' ],
             });
         } );
-
-        window.setTimeout(function() {
-            $("#alert").fadeTo(500, 0).slideUp(500, function(){
-                $(this).remove();
-            });
-        }, 5000);
     </script>
 @stop
